% Generated by roxygen2 (4.0.1): do not edit by hand
\name{autoLoad}
\alias{autoLoad}
\title{(do not use)}
\usage{
autoLoad()
}
\description{
Load all files in 'calc/autoload' - but this is now done with \code{fetchAll}. This function is kept for now to not break old code.
}
\author{
Henrik Renlund
}

