% Generated by roxygen2 (4.0.1): do not edit by hand
\name{branchTextFile}
\alias{branchTextFile}
\title{Create a information text file for a git branch}
\usage{
branchTextFile(name, file_name)
}
\arguments{
\item{name}{the name the project}

\item{file_name}{the file name to be used}
}
\description{
Create a information text file for a git branch
}

