% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cap_att.r
\name{cap_att}
\alias{cap_att}
\title{Caption for attaching}
\usage{
cap_att(caption, chunk)
}
\arguments{
\item{caption}{the caption}

\item{chunk}{the chunk that contains the plot}
}
\description{
Caption that possibly makes a caption that attaches a figure
}
\note{
This function needs a bit of work?
}

