% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cess.r
\name{cess}
\alias{cess}
\title{Run chunks}
\usage{
cess(chunk = NULL, file = NULL, envir = .GlobalEnv)
}
\arguments{
\item{chunk}{name or number of chunk, if \code{NULL} (default) then the
first chunk will be executed}

\item{file}{the file which has the chunks, default: 'rapport.rnw'}

\item{envir}{is \code{.GlobalEnv} by default, the environemnt in which to
evaluate chunks}
}
\description{
Run a chunk in a given document
}
\note{
This will replace \code{proh::first}
}

