% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_struc_functions.r
\name{chunks_info}
\alias{chunks_info}
\title{Chunks info}
\usage{
chunks_info(file = NULL, all = FALSE)
}
\arguments{
\item{file}{file of interest ( = \code{source_file} by default)}

\item{all}{if TRUE some extra chunk info will be given}
}
\value{
A data frame with variables
\itemize{
   \item name: name of chunk (if any)
   \item row: start row
   \item n.row: number of rows
   \item stop: end line of chunk
   \item eval.arg: (if \code{all = TRUE}) if there is an argument specified for \code{eval}
   \item code:  (if \code{all = TRUE}) the code in the chunk
}
}
\description{
Get information about chunks in a rnw file
}
