% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.r
\name{cmp}
\alias{cmp}
\alias{cmp_rmd}
\alias{cmp_rnw}
\title{Compile the rapport}
\usage{
cmp(input = NULL, ...)

cmp_rnw(input, clean = TRUE, look = FALSE, ...)

cmp_rmd(input, twice = FALSE, clean = FALSE, look = FALSE, ...)
}
\arguments{
\item{input}{should be 'rapport.rnw' (but can be changed)}

\item{...}{arguments to be passed on}

\item{clean}{should the LaTeX files be cleaned?}

\item{look}{should the pdf be opened after compilation?}

\item{twice}{need two code executions? (e.g. for counters)}

\item{clean}{should the LaTeX files be cleaned?}

\item{look}{should the pdf be opened after compilation?}
}
\description{
Compile the rapport
}
\section{Functions}{
\itemize{
\item \code{cmp_rnw}: Compile rnw files

\item \code{cmp_rmd}: Compile rmd files
}}
\author{
Henrik Renlund
}

