% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.r
\name{cmp_rnw}
\alias{cmp_rnw}
\alias{cmp}
\alias{cmp_dm}
\title{Compile rnw files}
\usage{
cmp_rnw(input, output, ...)

cmp(cess = TRUE, ...)

cmp_dm(cess = TRUE, ...)
}
\arguments{
\item{input}{file to compile}

\item{output}{output tex file}

\item{...}{arguments passed to \code{\link[knitr]{knit2pdf}}}

\item{cess}{run \code{proh::cess} to get (possible) defaults in first chunk?}
}
\description{
Wrapper for \code{knitr::knit2pdf}
}
\section{Functions}{
\itemize{
\item \code{cmp}: apply to 'source_file'

\item \code{cmp_dm}: apply to 'dm_source_file'
}}

