% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comp}
\alias{comp}
\title{Compile the rapport}
\usage{
comp(input = "rapport.rnw", settings = TRUE, clean = TRUE, look = FALSE,
  ...)
}
\arguments{
\item{input}{should be 'rapport.rnw' (but can be changed)}

\item{settings}{should the settings be read off '_META_.r'?}

\item{clean}{should the LaTeX files be cleaned?}

\item{look}{should the pdf be opened after compilation?}

\item{...}{arguments to be passed to \code{knit}}
}
\description{
Compile the rapport
}
\author{
Henrik Renlund
}

