% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetch}
\alias{fetch}
\title{Fetch}
\usage{
fetch(name, overwrite = TRUE, message = FALSE)
}
\arguments{
\item{name}{character; the name of a variable}

\item{overwrite}{if variable already exists in global workspace, should it be overwritten?}

\item{message}{do you want a message?}
}
\description{
Load an object from 'calc' directory
}
\author{
Henrik Renlund
}

