% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchAll}
\alias{fetchAll}
\title{Load all}
\usage{
fetchAll(calc = TRUE, autoload = TRUE)
}
\arguments{
\item{calc}{should contents of directory 'calc' be loaded?}

\item{autoload}{should contents of directory 'calc/autoload' be loaded?}
}
\description{
Load the contents of 'calc' and its subdirectories (in particular 'autoLoad')
}
\author{
Henrik Renlund
}

