% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figh.r
\name{figh}
\alias{figh}
\title{Caption generator}
\usage{
figh(s, attach_graph = NULL, ext = NULL, warn = TRUE)
}
\arguments{
\item{s}{Caption}

\item{attach_graph}{attach graph? if \code{NULL} then this parameter will
be taken from \code{opts_proh$get('attach_graph')}}

\item{ext}{what is the file extension? if \code{NULL} then this parameter will
be taken from \code{opts_proh$get('graph_dev')}}

\item{warn}{warn if extension does not seem available?}
}
\description{
Use this with chunk option \code{fig.cap} to attach a copy
of the produced graphic in the resulting pdf.
}
\note{
You might want to  specify something like
  \\DeclareGraphicsExtensions{.pdf,.eps,.png,.jpg,jpeg}
}

