% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileName.r
\name{file_name}
\alias{fileName}
\alias{file_name}
\title{Decompose filename}
\usage{
file_name(filename)

fileName(filename)
}
\arguments{
\item{filename}{the vector of names you want decomposed}
}
\value{
A dataframe with variables 'name' and 'extension'
}
\description{
Decompose a filename into name and extension
}
\section{Functions}{
\itemize{
\item \code{fileName}: An alias
}}
\examples{
some_names <- c("foo.txt", "foo.bar.r", ".emacs", "DESCRIPTION")
fileName(some_names)
}

