% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_length.r
\name{fixed_length}
\alias{fixed_length}
\title{fixed length}
\usage{
fixed_length(x, n = NULL, fill = " ", right = TRUE)
}
\arguments{
\item{x}{vector to fix}

\item{n}{uniformal nchar length for elements of x}

\item{fill}{character to fill the void}

\item{right}{fill to the right?}
}
\value{
character string
}
\description{
elements of uniformal length
}
\details{
make nchar have the same value for all elements of x
}
\examples{
fixed_length(x = c("A", "ABC"))
fixed_length(x = c("A", "ABC"), 2, fill = "_")
fixed_length(x = c("A", "ABC"), fill = ".", right = F)
fixed_length(8:12, fill = "0")
}
\author{
Henrik Renlund
}

