% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gitBranch.r
\name{gitBranch}
\alias{gitBranch}
\title{Create git branch}
\usage{
gitBranch(name)
}
\arguments{
\item{name}{Name of branch}
}
\description{
Create git branch. This function is mainly a helper for \code{send}.
}
\author{
Henrik Renlund
}

