% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keep}
\alias{keep}
\title{Save}
\usage{
keep(name, autoload = FALSE)
}
\arguments{
\item{name}{character; the name of one or more variable}

\item{autoload}{should the object be saved in subdirectory 'calc/autoload'?}
}
\description{
Save an object to 'calc' folder
}
\author{
Henrik Renlund
}

