% Generated by roxygen2 (4.0.2): do not edit by hand
\name{killCache}
\alias{killCache}
\title{Kill Cache}
\usage{
killCache(cache = TRUE, figure = TRUE, table = TRUE)
}
\arguments{
\item{cache}{remove files from sub directory 'cache'? (default: TRUE)}

\item{figure}{remove files from sub directory 'figure'? (default: TRUE)}

\item{table}{remove files from sub directory 'table'? (default: TRUE)}
}
\description{
Remove the contents of subdirectory 'cache', 'figure' and 'table'.
}
\details{
Sometimes it is easier clean the cache than to set \code{CACHE = FALSE} globally, since you may have forgotten to specify all dependencies...
}
\author{
Henrik Renlund
}

