% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/md_cap.r
\name{md_enumerate_at_ref}
\alias{md_enumerate_at_ref}
\title{Create enumeration of labels}
\usage{
md_enumerate_at_ref()
}
\description{
Create List of Tables/Figures and thus convert all labels
  created with \code{md_tcap} and \code{md_fcap} to numbers. Will most
  likely be removed in the near future.
}

