% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mumbo.R
\name{mumbo}
\alias{mumbo}
\title{random mumbo jumbo}
\usage{
mumbo(n, m = 2, try.unique = TRUE)
}
\arguments{
\item{n}{integer, how many random strings}

\item{m}{integer, length of each string}

\item{try.unique}{logical, make an honest try of getting the strings to be
unique?}
}
\description{
create random strings of length m
}
