% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/omni.r
\name{omni}
\alias{omni}
\title{Keep an object omnipresent}
\usage{
omni(name)
}
\arguments{
\item{name}{character; the name of a variable}
}
\description{
Shorthand for \code{keep(..., autoload=TRUE)}
}
\details{
If .rnw files are cached, sometimes the objects used in 'Sexpr' are
unavailable. 'omni' stores variables in 'calc/autoload' which are
automatically loaded in the first chunk of the rapport.
}
\author{
Henrik Renlund
}

