% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_loop_info}
\alias{print_loop_info}
\title{Print loop info}
\usage{
print_loop_info(i, n, len = 10)
}
\arguments{
\item{i}{loopin index (\code{i in 1:n})}

\item{n}{length of looping index}

\item{len}{number of times to get progress information}
}
\description{
Prints progress
}

