\docType{package}
\name{proh}
\alias{proh}
\alias{proh-package}
\title{proh.}
\description{
proh: pro(ject) h(enrik style)

This is a package for working with projects - which here means creating a (statistical) rapport with R, knitr and git version control. There are functions to create a certain directory structure and other very specific things. Note; these are tailored after the idiosyncratic preferences of the author.

If for some strange reason you are interested, read the documentation for \code{newProject} to get some idea. 
}

