% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm2.R
\name{rm2}
\alias{rm2}
\title{remove2}
\usage{
rm2(..., rmlist = NULL, envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{...}{variables you want to keep}

\item{rmlist}{alternatively, the variables you want to keep as a character vector}

\item{envir}{environment in which to find the variables}

\item{verbose}{give helful messages?}
}
\description{
remove all but
}
\details{
like \code{rm} but specify what you want to keep
}
