% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc_struc_functions.r
\name{sections_info}
\alias{sections_info}
\title{Sections info}
\usage{
sections_info(file = "rapport.rnw")
}
\arguments{
\item{file}{file of interest ( = 'rapport.rnw' by default)}
}
\value{
A data frame with variables
\itemize{
   \item name: name of section
   \item row: row number where section starts
   \item sub: number of 'sub', i.e. 0 for section, 1 for subsection,
   and 2 for subsubsection
}
}
\description{
Get information about sections in a rnw file
}

