% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.r
\name{send_function}
\alias{send}
\alias{send_dm}
\alias{send_function}
\title{mark file for delivery}
\usage{
send_function(output_file, version, source_file, git)

send(git = TRUE)

send_dm(git = TRUE)
}
\arguments{
\item{output_file}{the rapport pdf}

\item{version}{version number}

\item{source_file}{the rapport rnw}

\item{git}{commit? (default \code{TRUE})}
}
\description{
date and (possibly) version stamp a rapport and place in 'sent' folder
}
\section{Functions}{
\itemize{
\item \code{send}: use proh options 'output_file', 'version', and
'source_file' as defaults

\item \code{send_dm}: use proh options 'dm_output_file', 'dm_version', and
'dm_source_file' as defaults
}}

