% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shoWes.r
\name{shoWes}
\alias{shoWes}
\title{Available wesanderson palettes}
\usage{
shoWes()
}
\description{
Show available wesanderson palettes
}
\details{


Makes a plot visualizing the palettes from the wesanderson package
}

