% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tableh}
\alias{tableh}
\title{Create table with attached info}
\usage{
tableh(object, label, fun = write.csv, caption, data, ...)
}
\arguments{
\item{object}{usually a data frame or matrix}

\item{label}{a label 'tab:<label>' will be set in \code{latex}}

\item{fun}{what function to use to write \code{object} to file 'table/<label>.txt'.}

\item{caption}{caption for \code{latex}}

\item{data}{if attach data is something other than object}

\item{...}{arguments passed to \code{latex}}
}
\description{
\code{Hmisc::latex} and object, write to file and attach
}

