% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toDate}
\alias{toDate}
\title{Transform to date}
\usage{
toDate(x, not_dates = c("", " "), ...)
}
\arguments{
\item{x}{character or factor}

\item{not_dates}{the things that should be NA}

\item{...}{arguments passed to \code{as.Date}}
}
\description{
wrapper for \code{as.Date} that replaces ""-strings with
\code{NA} before conversion
}

