% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mumbo.R
\name{urnames}
\alias{urnames}
\title{Rownames generator for LaTeX tables}
\usage{
urnames(df, x = NULL)
}
\arguments{
\item{df}{data frame}

\item{x}{rownames to be visible, in particular non-unique such names}
}
\description{
Make ostensibly (in LaTeX tables) non-unique rownames by adding random
    letters hidden in 'phantom' code. This is useful in combination with
    'rgroups'-feature of \code{Hmisc::latex}
}
\examples{
df <- data.frame(x = 1:100)
rownames(df) <- urnames(df, rep("A", 100))
}

