% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cap_att}
\alias{cap_att}
\title{Caption for attaching}
\usage{
cap_att(caption, chunk)
}
\arguments{
\item{caption}{the caption}

\item{chunk}{the chunk that contains the plot}
}
\description{
Caption that possible makes a caption that attaches a figure
}

