% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.r
\name{cmp}
\alias{cmp}
\alias{cmp_rnw}
\title{Compile the rapport}
\usage{
cmp(input = NULL, cess = TRUE, ...)

cmp_rnw(input, clean = TRUE, look = FALSE, ...)
}
\arguments{
\item{input}{file to be compiled}

\item{cess}{run proh::cess to get (possible) defaults in first chunk?}

\item{...}{arguments to be passed on}

\item{clean}{should the LaTeX files be cleaned?}

\item{look}{should the pdf be opened after compilation?}
}
\description{
Compile the rapport
}
\section{Functions}{
\itemize{
\item \code{cmp_rnw}: Compile rnw files
}}
\author{
Henrik Renlund
}

