% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc_struc_functions.r
\name{doc_struc}
\alias{doc_struc}
\title{Document structure}
\usage{
doc_struc(file = NULL)
}
\arguments{
\item{file}{file of interest ( = 'rapport.rnw' by default)}
}
\value{
A print out
}
\description{
Get information about the structure in a rnw file
}

