% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.r
\name{fetch}
\alias{fetch}
\title{Fetch}
\usage{
fetch(name, overwrite = TRUE, message = FALSE, autoload = FALSE,
  formats = c("rdata", "rdat"), env = .GlobalEnv)
}
\arguments{
\item{name}{character; the name of a variable. If missing, available objects will be listed}

\item{overwrite}{if variable already exists in global workspace, should it be overwritten?}

\item{message}{do you want an explanatory message?}

\item{autoload}{is the variable in subdirectory 'autoload'? Default \code{FALSE}}

\item{formats}{formats to look for. Default \code{c('.rdata', '.rdat')}.}

\item{env}{the environment to load into. Defaults to the global enviroment.}
}
\description{
Load an object from 'calc' directory
}
\author{
Henrik Renlund
}

