% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.r
\name{fetch}
\alias{fetch}
\alias{fetch_}
\title{Fetch}
\usage{
fetch(name = NULL, overwrite = FALSE, message = FALSE, autoload = FALSE,
  formats = c("rdata", "rdat", "Rdata", "Rdat"), env = .GlobalEnv)

fetch_(..., overwrite = TRUE, message = FALSE, autoload = FALSE,
  formats = c("rdata", "rdat"), env = .GlobalEnv)
}
\arguments{
\item{name}{character; the name of a variable. If missing, available objects will be listed}

\item{overwrite}{if variable already exists in global workspace, should it be overwritten?}

\item{message}{do you want an explanatory message?}

\item{autoload}{is the variable in subdirectory 'autoload'? Default \code{FALSE}}

\item{formats}{formats to look for. Default \code{c('.rdata', '.rdat')}.}

\item{env}{the environment to load into. Defaults to the global enviroment.}

\item{...}{(possibly unquoted) names of objects}
}
\description{
Load an object from 'calc' directory
}
\section{Functions}{
\itemize{
\item \code{fetch_}: Non-standard evaluation version
}}
\author{
Henrik Renlund
}

