% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fileName.r
\name{fileName}
\alias{fileName}
\title{Decompose filename}
\usage{
fileName(filename)
}
\arguments{
\item{filename}{the vector of names you want decomposed}
}
\value{
A dataframe with variables 'name' and 'extension'
}
\description{
Decompose a filename into name and extension
}
\examples{
some_names <- c("foo.txt", "foo.bar.r", ".emacs", "DESCRIPTION")
fileName(some_names)
}

