% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imposeProject}
\alias{imposeProject}
\title{Impose project structure}
\usage{
imposeProject(path = NULL, class = "ucr", go_there = TRUE,
  RSproj = TRUE, git = TRUE)
}
\arguments{
\item{path}{Path to project directory (else current)}

\item{class}{Class of document in 'rapport.rnw' (default: 'ucr')}

\item{go_there}{Set working directory to project directory? (default: TRUE)}

\item{RSproj}{Start a RStudio project? (deault: TRUE)}

\item{git}{should git be initialized? (also a .gitignore file will be created)}
}
\description{
Impose project structure in existing directory, files and directories created by \code{newProject} will be
}
\author{
Henrik Renlund
}
\seealso{
\code{proh::newProject}
}

