% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.r
\name{keep}
\alias{keep}
\alias{keep_}
\title{Save}
\usage{
keep(name, autoload = FALSE)

keep_(..., autoload = FALSE)
}
\arguments{
\item{name}{character; the name of one or more variable}

\item{autoload}{should the object be saved in subdirectory 'calc/autoload'?}

\item{...}{(possibly unquoted) names of objects}
}
\description{
Save an object to 'calc' folder
}
\section{Functions}{
\itemize{
\item \code{keep_}: Non-standard evaluation version
}}

\author{
Henrik Renlund
}
