% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.r
\name{look}
\alias{look}
\title{View the pdf}
\usage{
look(file = NULL)
}
\arguments{
\item{file}{filename (default 'rapport.pdf')}
}
\description{
Open the current pdf version of the rapport (or indeed try to
open any file, as this is just a wrapper for \code{shell.exec})
}
\author{
Henrik Renlund
}

