% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/look.r
\name{look}
\alias{look}
\title{View the pdf}
\usage{
look(sumatra = FALSE)
}
\arguments{
\item{sumatra}{not implemented yet, will be FALSE}
}
\description{
Open the current pdf version of the rapport
}
\author{
Henrik Renlund
}

