% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/md_cap.r
\name{md_cap}
\alias{md_cap}
\alias{md_fcap}
\alias{md_tcap}
\title{Capions in rmarkdown}
\usage{
md_cap(caption, label, type = NULL, at.ref = FALSE)

md_tcap(caption = "", label = NULL, at.ref = FALSE)

md_fcap(caption = "", label = NULL, at.ref = FALSE)
}
\arguments{
\item{caption}{the caption}

\item{label}{the label you can refer to in the .rmd document. If \code{NULL}
it will inherit the 'label' from the current chunk}

\item{type}{"table" or "figure". User do not have to care as you use
\code{md_tcap} or \code{md_fcap}, which calls \code{md_cap} with this
argument fixed}

\item{at.ref}{if \code{TRUE} use pandoc markdown's enumerated list to create
the references (default \code{FALSE}). This option will most likely be
removed in the future!}
}
\description{
This function allows enumerated tables and figures in rmarkdown
  with html or word output (and pdf I suppose).
}
\details{
The \code{md_cap} function adds a list element to the \code{loft}
  list in \code{opts_proh}. Use \code{md_tcap} for tables and
  \code{md_fcap}. (If \code{at.ref} is \code{TRUE} these are converted by
  \code{md_enumerate_at_ref} to two referrable, numbered list (List of Tables
  and List of Figures, respectively), which, when created, will have the
  effect of replacing the @-label with a number.
}
\section{Functions}{
\itemize{
\item \code{md_tcap}: Captions for tables

\item \code{md_fcap}: Captions for figures
}}
\note{
If the chunk with \code{md_enumerate_labels} is at the beginning of the
  document you probably have to \code{knit} the document twice before the
  List of Tables/Figures appears, since the \code{md_cap} function has not
  actually been called yet.
}
\examples{
# <<>>=
# md_enumerate_labels() # knit twice first time
# @
#
# Below is Table @myLabel.
#
# <<>>=
# object <- data.frame("Foo" = 1:2, "Bar" = letters[1:2])
# knitr::kable(object, caption = md_tcap("myLabel", "My Caption))
# @
}

