% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omni.r
\name{omni}
\alias{omni}
\alias{omni_}
\title{Keep an object omnipresent}
\usage{
omni(name)

omni_(...)
}
\arguments{
\item{name}{character; the name of a variable}

\item{...}{(possibly unquoted) names of objects}
}
\description{
Shorthand for \code{keep(..., autoload=TRUE)}
}
\details{
If .rnw files are cached, sometimes the objects used in 'Sexpr' are
unavailable. 'omni' stores variables in 'calc/autoload' which are
automatically loaded in the first chunk of the rapport.
}
\section{Functions}{
\itemize{
\item \code{omni_}: Non-standard evaluation version
}}
\author{
Henrik Renlund
}

