% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{opts_proh}
\alias{opts_proh}
\title{opts (not useful yet)}
\format{\preformatted{List of 3
 $ get    :function (name)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 21 13 32 1 13 1 21 32
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000012949950> 
 $ set    :function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 38 13 46 1 13 1 38 46
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000012949950> 
 $ restore:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 51 12 59 1 12 1 51 59
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000012949950> 
}}
\usage{
opts_proh
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently two values are maintained with the functions in (the list) opts_proh: \itemize{
\item attach_graph
\item attach_table
}
Currently, these values has no effect on anything - but hopefully they will
in the future...
}
\keyword{datasets}

