% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.r
\docType{data}
\name{opts_proh}
\alias{opts_proh}
\title{proh options}
\format{An object of class \code{list} of length 4.}
\usage{
opts_proh
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_proh}:
\itemize{
\item main_document - default: rapport.rnw
\item output_format - default: pdf_document
\item output_file - will be like main_document but appropriate file extension (
unless set manually)
\item attach_graph - use \code{figh} in \code{fig.caption}
(in a \code{knitr} chunk), i.e. \code{fig.caption = figh("My Caption")},
this forces \code{figh} to include an attach-statement
pointing to the relevant figure in 'figure/' .
\item graph_dev - graphical extension
\item attach_table - this forces \code{tableh} to also attach its tables
\item table_fnc - this is the function to write the \code{tableh}-table
to file
}
}
\keyword{datasets}

