% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.r
\docType{data}
\name{opts_proh}
\alias{opts_proh}
\title{proh options}
\format{\preformatted{List of 4
 $ get    :function (name)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 13 26 1 13 1 14 26
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000014e94fc0> 
 $ set    :function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 32 13 42 1 13 1 32 42
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000014e94fc0> 
 $ restore:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 47 17 61 1 17 1 47 61
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000014e94fc0> 
 $ check  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 66 15 101 1 15 1 66 101
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x0000000014e94fc0> 
}}
\usage{
opts_proh
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_proh}:
\itemize{
\item main_document - default: rapport.rnw
\item output_format - default: pdf_document
\item output_file - will be like main_document but appropriate file extension (
unless set manually)
\item attach_graph - use \code{figh} in \code{fig.caption}
(in a \code{knitr} chunk), i.e. \code{fig.caption = figh("My Caption")},
this forces \code{figh} to include an attach-statement
pointing to the relevant figure in 'figure/' .
\item graph_dev - graphical extension
\item attach_table - this forces \code{tableh} to also attach its tables
\item table_fnc - this is the function to write the \code{tableh}-table
to file
}
}
\keyword{datasets}

