% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.r
\docType{data}
\name{opts_proh}
\alias{opts_proh}
\title{proh options}
\format{An object of class \code{list} of length 4.}
\usage{
opts_proh
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_proh}:
\itemize{
\item source_file - default: rapport.rnw
\item output_file - will be like source_file but appropriate file extension (
unless set manually)
\item version a version number as character, e.g. "Version 1". This will
    appear on the LaTeX version of the rapport
\item version_latex this is the string that will determine how the version
    number appears in the title (in LaTeX produced pdf:s) and will be set
    automatically
}
}
\keyword{datasets}

