% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_loop_info.r
\name{print_loop_info}
\alias{print_loop_info}
\title{Print loop info}
\usage{
print_loop_info(i, n, len = 10, prefix = "", suffix = "")
}
\arguments{
\item{i}{loopin index (\code{i in 1:n})}

\item{n}{length of looping index}

\item{len}{number of times to get progress information}

\item{prefix}{printing prefix}

\item{suffix}{printing suffix}
}
\description{
Prints progress
}
\examples{
n <- 1E3
for(i in 1:n){
   print_loop_info(i, n)
   print_loop_info(i, n, len = 20, "       ", " <yeah>")
}; rm(i, n)
}

