\docType{package}
\name{proh}
\alias{proh}
\alias{proh-package}
\title{proh.}
\description{
This is a package for working with projects - which for this purpose means creating a (statistical) rapport with R, knitr and git version control. There are functions to create a certain directory structure and functions to do some very specific things. Note; most likely very idiosyncratic to  the author.

If for some strange reason you are interested, read the documentation for \code{newProject} to get some idea. 
}

