% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tableh.r
\name{tableh}
\alias{tableh}
\title{Create table with attached info}
\usage{
tableh(object, attach_table, label, fun = write.csv, caption, data,
  fe = "txt", ...)
}
\arguments{
\item{object}{usually a data frame or matrix}

\item{attach_table}{logical; should the table be attached? if missing this
will be determined by \code{opts_proh$get("attach_table")}}

\item{label}{a label 'tab:<label>' will be set in \code{latex}}

\item{fun}{what function to use to write \code{object} to file
'table/<label>.txt'.}

\item{caption}{caption for \code{latex}}

\item{data}{if attach data is something other than object}

\item{fe}{file extension given if data is written to file}

\item{...}{arguments passed to \code{latex}}
}
\description{
\code{Hmisc::latex} and object, write to file and attach
}

