% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableh.r
\name{tableh}
\alias{tableh}
\title{Create table with attached info}
\usage{
tableh(object, attach_table = NULL, label = NULL, file = "",
  where = "htb", table_fnc = NULL, caption = NULL, caption.lot = NULL,
  data = object, fe = "txt", data.file = NULL, table.dir = "tableh",
  ...)
}
\arguments{
\item{object}{passed to \code{Hmisc::latex}}

\item{attach_table}{logical; should the table be attached? if missing this
will be determined by \code{opts_proh$get("attach_table")}}

\item{label}{passed to \code{Hmisc::latex}}

\item{file}{passed to \code{Hmisc::latex}}

\item{where}{passed to \code{Hmisc::latex}}

\item{table_fnc}{what function to use to write \code{object} to file.
If \code{NULL} (default) this will be taken from
\code{opts_proh$get("table_fnc")}}

\item{caption}{passed to \code{Hmisc::latex}}

\item{caption.lot}{passed to \code{Hmisc::latex}}

\item{data}{if attach data is something other than object}

\item{fe}{file extension given if data is written to file}

\item{data.file}{name of file 'object' or 'data' is written to}

\item{table.dir}{directory name in which to store 'data.file'}

\item{...}{arguments passed to \code{Hmisc::latex}}
}
\description{
\code{Hmisc::latex} and object, write to file and attach
}

